Commit 1.0
Set the initial page to be the login page. Added a username and password box, as well as a currently nonfunctional login button (on account of not having another place to link to) and a place to put a link to the account registration page. Changed some CSS stuff for the id "header_div" for formatting.
It's a start, I guess? I'll do some more tomorrow, if not later today